// modules/worker/sendEmailByCollaborator.Type.ts
export type SettlementEmailType = {
    id: string;
    recipitentName: string;
    email: string;
    active: boolean;
    mobileNumber: string;
    identification: string;
    position: string;
    collaboratorId: string;
    collaboratorIdentification: string;
    name: string;
    bp: string;
    createdAt: Date;
    updatedAt: Date;
};